Added Scene for the second chapter ,fix TextMeshPro issues,renamed scenes and created script to show models based on the page of the text.
Created new scene for the second chapter.Edited pictures and splited in into sprites for showing.Created animation in order for the complete picture to appear in the end with the scrolling to right motion.
Fixed some issues with the TextMeshPro where i kept not rendering some Greek Alphabet characters.
Renamed the chapters to be shown in correct order.
Created a script to show and hide models based on the page of the shown text.
Added some more 3d models