Added scripts,edited scenes and made animations
Added scripts so that the 3d models can be moved and rotated in the screen with touch instead of moving the image.
Added extra 3d models
Made animation of swords clashing and edited the map scenes.
Fixed some issues with the next and back page buttons and that the models will appear and disappear based on the page that is about them
Added an outro text of what was learned
Fixed TechMeshPro issue with character rending.